Split EngineRouter.swift into handler groups and drop shadowed copies (#1174)
makeEngineRouter is now a thin assembler that merges per-concern
Engine*RPCHandlers.swift maps. Daemon-owned method names are no longer
registered on the engine, so a name has one body. The parity script globs
every engine handler file and fails if a shadowed copy returns.

Crow-Session: C4CD06A1-40F7-47B2-A533-C2CC84157EFB
