+++ scripts/check-cli-parity.sh
@@ -10,9 +10,13 @@
 #      The glob is load-bearing: a group that doesn't match it is invisible to
 #      this gate. Support files — `*RPCSupport.swift`, `RPCWebSocketHandler.swift`,
 #      `RPCDispatcher.swift`, `RPCLanePolicy.swift` — deliberately do not match.)
-#   - engine handlers → Packages/CrowEngine/Sources/CrowEngine/EngineRouter.swift
-#     (the daemon router is built with `fallback: makeEngineRouter(ctx)`, so the
-#      live surface is the union of the two dictionaries)
+#   - engine handlers → Packages/CrowEngine/Sources/CrowEngine/*RPCHandlers.swift
+#     plus EngineRouter.swift (the assembler). The daemon router is built with
+#     `fallback: makeEngineRouter(ctx)`, so the live surface is the union of the
+#     two dictionaries. The engine glob is load-bearing the same way as the
+#     daemon's: after CROW-1174 the assembler has no `"method": {` keys, so a
+#     stale glob that only sees EngineRouter.swift would under-count live
+#     fallback methods (`hook-event`, `send`, `new-terminal`, …).
 #   - the ledger      → Packages/CrowCore/Sources/CrowCore/Parity/ParityLedger.swift
 #
 # `CrowDaemonTests/RPCLedgerParityTests` makes the same assertion against the
@@ -24,14 +28,19 @@ set -euo pipefail
 
 ROOT="$(cd "$(dirname "${BASH_SOURCE[0]}")/.." && pwd)"
 DAEMON_DIR="$ROOT/Packages/CrowDaemon/Sources/CrowDaemon"
-ENGINE="$ROOT/Packages/CrowEngine/Sources/CrowEngine/EngineRouter.swift"
+ENGINE_DIR="$ROOT/Packages/CrowEngine/Sources/CrowEngine"
+ENGINE_ASSEMBLER="$ENGINE_DIR/EngineRouter.swift"
 LEDGER="$ROOT/Packages/CrowCore/Sources/CrowCore/Parity/ParityLedger.swift"
 
 if [ ! -d "$DAEMON_DIR" ]; then
   echo "check-cli-parity: missing $DAEMON_DIR" >&2
   exit 1
 fi
-for f in "$ENGINE" "$LEDGER"; do
+if [ ! -d "$ENGINE_DIR" ]; then
+  echo "check-cli-parity: missing $ENGINE_DIR" >&2
+  exit 1
+fi
+for f in "$ENGINE_ASSEMBLER" "$LEDGER"; do
   if [ ! -f "$f" ]; then
     echo "check-cli-parity: missing $f" >&2
     exit 1
@@ -43,16 +52,30 @@ done
 # treat as a missing file with a confusing error.
 shopt -s nullglob
 DAEMON_FILES=("$DAEMON_DIR"/*RPCHandlers.swift)
+ENGINE_FILES=("$ENGINE_DIR"/*RPCHandlers.swift)
 shopt -u nullglob
 
-# Vacuity: if the glob stops matching the extracted groups, EngineRouter.swift
-# alone still has enough methods to clear MIN_METHODS below, and the ledger
-# comparison would silently ignore every daemon verb. Fail loudly instead.
+# Vacuity: if the glob stops matching the extracted groups, the engine
+# fallback files plus an empty assembler would under-count (or after
+# CROW-1174, fail MIN_METHODS). Fail with a specific glob error instead.
 MIN_DAEMON_FILES=8
 if [ "${#DAEMON_FILES[@]}" -lt "$MIN_DAEMON_FILES" ]; then
   echo "check-cli-parity: only ${#DAEMON_FILES[@]} *RPCHandlers.swift files under $DAEMON_DIR (expected >= $MIN_DAEMON_FILES)." >&2
   echo "The daemon-handler glob has probably gone stale — registrations live in" >&2
   echo "Packages/CrowDaemon/Sources/CrowDaemon/*RPCHandlers.swift (CROW-1134)." >&2
+  exit 1
+fi
+
+# Same floor for the engine fallback groups. After CROW-1174 the assembler
+# (`EngineRouter.swift`) has no handler keys; a glob that misses
+# Engine*RPCHandlers.swift would under-count live fallback methods and the
+# ledger comparison would go green while crowd lost `hook-event` / `send` /
+# `new-terminal`.
+MIN_ENGINE_FILES=4
+if [ "${#ENGINE_FILES[@]}" -lt "$MIN_ENGINE_FILES" ]; then
+  echo "check-cli-parity: only ${#ENGINE_FILES[@]} *RPCHandlers.swift files under $ENGINE_DIR (expected >= $MIN_ENGINE_FILES)." >&2
+  echo "The engine-handler glob has probably gone stale — registrations live in" >&2
+  echo "Packages/CrowEngine/Sources/CrowEngine/*RPCHandlers.swift (CROW-1174)." >&2
   exit 1
 fi
 
@@ -67,10 +90,32 @@ fi
 HANDLER_RE='^[[:space:]]+"[a-z0-9-]+": \{ (@Sendable )?[A-Za-z_][A-Za-z0-9_]* in([[:space:]]|$)'
 
 registered_methods() {
-  grep -hoE "$HANDLER_RE" "${DAEMON_FILES[@]}" "$ENGINE" \
+  grep -hoE "$HANDLER_RE" "${DAEMON_FILES[@]}" "${ENGINE_FILES[@]}" "$ENGINE_ASSEMBLER" \
     | sed -E 's/^[[:space:]]*"([a-z0-9-]+)".*/\1/' \
     | sort -u
 }
+
+# CROW-1174: a name must exist in at most one of the daemon maps or the engine
+# maps. Re-introducing a shadowed copy would not change the ledger union, so
+# this is a separate gate — the drift ADR 0016 recorded (`get-config`
+# `app_running: true` vs `false`) was exactly this failure mode.
+daemon_methods() {
+  grep -hoE "$HANDLER_RE" "${DAEMON_FILES[@]}" \
+    | sed -E 's/^[[:space:]]*"([a-z0-9-]+)".*/\1/' \
+    | sort -u
+}
+engine_methods() {
+  grep -hoE "$HANDLER_RE" "${ENGINE_FILES[@]}" "$ENGINE_ASSEMBLER" \
+    | sed -E 's/^[[:space:]]*"([a-z0-9-]+)".*/\1/' \
+    | sort -u
+}
+overlap="$(comm -12 <(daemon_methods) <(engine_methods) || true)"
+if [ -n "$overlap" ]; then
+  echo "check-cli-parity: engine files re-register methods the daemon already owns:" >&2
+  echo "$overlap" | sed 's/^/  /' >&2
+  echo "Delete the engine copy — crowd answers the daemon body via fallback." >&2
+  exit 1
+fi
 
 # Ledger rows are `.read("method", …)` / `.write("method", …)`. A row with a long
 # exemption reason wraps, putting the method literal on its own line, so the
